[ADD] Timer View
Timer 기본 화면 추가

+++ domain/src/main/java/kr/hs/dgsw/domain/repository/TimerRepository.kt
@@ -3,5 +3,5 @@ package kr.hs.dgsw.domain.repository
 import io.reactivex.Observable
 
 interface TimerRepository {
-    fun countDownTime(partTime: Long): Observable<Int>
+    fun countDownTime(partTime: Int): Observable<Int>
 }
